# docs/swagger.yaml
basePath: /
definitions:
  apierror.APIError:
    properties:
      code:
        type: integer
      sysMessage:
        type: string
      timestamp:
        type: string
      userMessage:
        type: string
    type: object
...
  domain.Template:
    properties:
      content:
        type: string
      created_at:
        type: string
      fields:
        items:
          type: string
        type: array
      id:
        type: string
      is_html:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  domain.TemplateMetadata:
    properties:
      created_at:
        type: string
      fields:
        items:
          type: string
        type: array
      id:
        type: string
      is_html:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.CreateTemplateResponse:
    properties:
      id:
        example: c790g02f8n90184b23qg
        type: string
    type: object
...
  handlers.GetTemplateRawResponse:
    properties:
      template:
        $ref: '#/definitions/domain.Template'
    type: object
...
  handlers.ListTemplateMetadataResponse:
    properties:
      templates:
        items:
          $ref: '#/definitions/domain.TemplateMetadata'
        type: array
    type: object
...
  handlers.UpdateTemplateResponse:
    properties:
      id:
        example: c790g02f8n90184b23qg
        type: string
    type: object
host: localhost:2500
info:
  contact: {}
  description: Carrpigeon email distribution and template service API.
  title: Carrpigeon API
  version: "1.0"
paths:
...
  /templates:
    get:
      description: Get a paginated list of email templates metadata
      parameters:
      - default: 10
        description: Number of items to return (default 10)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListTemplateMetadataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: List templates metadata
      tags:
      - templates
    post:
      consumes:
      - multipart/form-data
      description: Uploads and saves a new HTML or plain text template file
      parameters:
      - description: Template name
        in: formData
        name: name
        required: true
        type: string
      - description: Template file (.html or .txt)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handlers.CreateTemplateResponse'
        "400":
          description: Bad Request or missing fields
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Upload template
      tags:
      - templates
  /templates/{id}:
    delete:
      description: Deletes an email template by ID
      parameters:
      - description: Template ID (xid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Template Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Delete template
      tags:
      - templates
    get:
      description: Get raw template content and metadata by ID
      parameters:
      - description: Template ID (xid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetTemplateRawResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Template Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Get raw template
      tags:
      - templates
    put:
      consumes:
      - multipart/form-data
      description: Updates an existing template's name and file content
      parameters:
      - description: Template ID (xid)
        in: path
        name: id
        required: true
        type: string
      - description: Template name
        in: formData
        name: name
        required: true
        type: string
      - description: Template file (.html or .txt)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UpdateTemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Template Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Update template
      tags:
      - templates
swagger: "2.0"